Check for .nomedia file inside tv show folders
Previousy, only the base source folder was checked for a .nomedia file,
which was rather pointless since you'd simply not add a source if you
didn't want it in your library.

Now the actual TV show folders are also checked, so you can add a
`.nomedia` file inside the folder of a TV show to exclude it from being
added to the library.